+++ kernel/memory/virtual/paging.h
@@ -23,7 +23,13 @@ struct PageTable {
     PageTableEntry entries[512];
 } __attribute__((packed));
 
-void map_pages(const uint64_t virtualAddress, const uint64_t physicalAddress, const uint8_t flags, const uint64_t requiredSize);
-void map_page(const uint64_t virtualAddress, const uint64_t physicalAddress, const uint8_t flags);
+#define PAGE_FLAG_WRITE 0b00000010
+#define PAGE_FLAG_CACHE_DISABLE 0b00010000
+
+bool map_page(const uint64_t virtualAddress, const uint64_t physicalAddress, const uint8_t flags);
+bool map_pages(const uint64_t virtualAddress, const uint64_t physicalAddress, const uint8_t flags, const uint64_t requiredSize);
+
+bool free_page(const uint64_t virtualAddress);
+bool free_pages(const uint64_t virtualAddress, const uint64_t requiredSize);
 
 #endif // PAGING_H